fix build error

+++ iocore/cache/CacheWrite.cc
@@ -107,8 +107,8 @@ CacheVC::updateVector(int /* event ATS_UNUSED */, Event * /* e ATS_UNUSED */)
       */
       if (alternate_index >= 0) {
         /*if length changed, fragment offset also need to change, if not cause range request error*/
-        if (total_len != write_vector->get(alternate_index)->object_size_get()) {
-          Debug("cache_update", "update length changed, new length:%d, old length:%d", total_len, write_vector->get(alternate_index)->object_size_get());  
+        if (total_len != uint64_t(write_vector->get(alternate_index)->object_size_get())) {
+          Debug("cache_update", "update length changed, new length:%lu, old length:%ld", total_len, write_vector->get(alternate_index)->object_size_get());  
         } else {
           /*do not copy everytimes, if data changed, this copy makes range request error*/
           alternate.copy_frag_offsets_from(write_vector->get(alternate_index));
